Move out cubic spline code to separate files.
